Initial commit of the Python2 script
This needs support software provided by Adafruit to go with the PCA9685 Servo Board that I'm using with my pi.
Robocorder is the main file. It parses the input and executes the finger switches. 
NewYork.txt is a file that shows the input format for the text files for the roborecorder to read. I haven't programmed it to play 'Welcome to New  York', yet. It's just placeholders now.